# backend/opstrack-api-spec.json
{"openapi":"3.0.0","paths":{"/api/assets":{"get":{"operationId":"AssetsController_findAll","summary":"Get all assets with optional filtering","parameters":[{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assets"]},"post":{"operationId":"AssetsController_create","summary":"Create a new physical equipment record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssetDto"}}}},"responses":{"201":{"description":""}},"tags":["Assets"]}},"/api/assets/search":{"get":{"operationId":"AssetsController_search","summary":"Full-Text Search across assets using PostgreSQL tsvector","parameters":[{"name":"q","required":true,"in":"query","description":"Search query keyword","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assets"]}},"/api/assets/{id}":{"get":{"operationId":"AssetsController_findOne","summary":"Get asset details including maintenance schedules & work orders","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assets"]},"delete":{"operationId":"AssetsController_remove","summary":"Delete an asset record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Assets"]}},"/api/work-orders":{"get":{"operationId":"WorkOrdersController_findAll","summary":"List all work orders with filters","parameters":[{"name":"status","required":false,"in":"query","schema":{"type":"string"}},{"name":"priority","required":false,"in":"query","schema":{"type":"string"}},{"name":"technicianId","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Work Orders"]},"post":{"operationId":"WorkOrdersController_create","summary":"Create a new maintenance work order","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["Work Orders"]}},"/api/work-orders/kanban":{"get":{"operationId":"WorkOrdersController_getKanbanBoard","summary":"Get work orders grouped for Angular Kanban board","parameters":[],"responses":{"200":{"description":""}},"tags":["Work Orders"]}},"/api/work-orders/{id}":{"get":{"operationId":"WorkOrdersController_findOne","summary":"Get single work order with asset, technician, and parts","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Work Orders"]}},"/api/work-orders/{id}/status":{"patch":{"operationId":"WorkOrdersController_updateStatus","summary":"Update work order status (e.g. dragging across Kanban columns)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusDto"}}}},"responses":{"200":{"description":""}},"tags":["Work Orders"]}},"/api/work-orders/{id}/complete":{"post":{"operationId":"WorkOrdersController_complete","summary":"Complete work order and atomically deduct consumed spare parts","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompleteWorkOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["Work Orders"]}},"/api/schedules":{"get":{"operationId":"SchedulesController_findAll","summary":"List all recurring preventative maintenance schedules","parameters":[{"name":"dueSoonOnly","required":false,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"tags":["Maintenance Schedules"]},"post":{"operationId":"SchedulesController_create","summary":"Create a new maintenance schedule rule","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateScheduleDto"}}}},"responses":{"201":{"description":""}},"tags":["Maintenance Schedules"]}},"/api/schedules/trigger-due-check":{"post":{"operationId":"SchedulesController_triggerCheck","summary":"Manually trigger the preventative maintenance evaluation worker","parameters":[],"responses":{"201":{"description":""}},"tags":["Maintenance Schedules"]}},"/api/parts":{"get":{"operationId":"SparePartsController_findAll","summary":"List spare parts inventory with optional low-stock alert filter","parameters":[{"name":"lowStockOnly","required":false,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"tags":["Spare Parts"]},"post":{"operationId":"SparePartsController_create","summary":"Add a new spare part to inventory","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePartDto"}}}},"responses":{"201":{"description":""}},"tags":["Spare Parts"]}},"/api/parts/{id}":{"get":{"operationId":"SparePartsController_findOne","summary":"Get single spare part details","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Spare Parts"]}},"/api/technicians":{"get":{"operationId":"TechniciansController_findAll","summary":"List all active technicians with active work order count","parameters":[],"responses":{"200":{"description":""}},"tags":["Technicians"]}},"/api/technicians/{id}":{"get":{"operationId":"TechniciansController_findOne","summary":"Get single technician profile and assigned work orders","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Technicians"]}}},"info":{"title":"OpsTrack CMMS API","description":"Enterprise Computerized Maintenance Management System API specification","version":"1.0","contact":{}},"tags":[{"name":"Assets","description":"Physical equipment and machinery lifecycle management"},{"name":"Work Orders","description":"Maintenance tasks, Kanban board, and stock consumption"},{"name":"Maintenance Schedules","description":"Preventative maintenance rules and automated triggers"},{"name":"Spare Parts","description":"Inventory levels, pricing, and threshold alerts"},{"name":"Technicians","description":"Technician profiles and active assignment metrics"}],"servers":[],"components":{"schemas":{"CreateAssetDto":{"type":"object","properties":{"name":{"type":"string","example":"Industrial Air Compressor X500"},"serial_number":{"type":"string","example":"AC-X500-8841"},"category":{"type":"string","example":"Pneumatics"},"location":{"type":"string","example":"Building A - Utility Room 2"},"status":{"type":"string","enum":["OPERATIONAL","UNDER_MAINTENANCE","DECOMMISSIONED","IN_STORAGE"],"default":"OPERATIONAL"},"purchase_date":{"type":"string","example":"2023-01-15"},"warranty_expires_at":{"type":"string","example":"2026-01-15"},"specs":{"type":"object","example":{"max_psi":175,"motor_hp":15}}},"required":["name","serial_number","category","location"]},"CreateWorkOrderDto":{"type":"object","properties":{"title":{"type":"string","example":"Emergency Hydraulic Pressure Drop Repair"},"description":{"type":"string","example":"Technician reported fluid seepage on primary piston cylinder."},"priority":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"],"default":"MEDIUM"},"status":{"type":"string","enum":["PENDING","IN_PROGRESS","COMPLETED","CANCELLED"],"default":"PENDING"},"asset_id":{"type":"string","example":"bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb"},"assigned_technician_id":{"type":"string","example":"22222222-2222-2222-2222-222222222222"},"schedule_id":{"type":"string","example":"90000002-0000-0000-0000-000000000002"},"scheduled_date":{"type":"string","example":"2026-09-15"}},"required":["title","asset_id"]},"UpdateStatusDto":{"type":"object","properties":{"status":{"type":"string","enum":["PENDING","IN_PROGRESS","COMPLETED","CANCELLED"],"example":"IN_PROGRESS"}},"required":["status"]},"ConsumedPartDto":{"type":"object","properties":{"part_id":{"type":"string","example":"55555555-5555-5555-5555-555555555552"},"quantity_used":{"type":"number","example":1}},"required":["part_id","quantity_used"]},"CompleteWorkOrderDto":{"type":"object","properties":{"parts_used":{"type":"array","items":{"$ref":"#/components/schemas/ConsumedPartDto"}}}},"CreateScheduleDto":{"type":"object","properties":{"asset_id":{"type":"string","example":"aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"},"task_name":{"type":"string","example":"Oil & Air Filter Replacement"},"frequency_interval_days":{"type":"number","example":90,"description":"Frequency interval in days"},"next_due_date":{"type":"string","example":"2026-10-01"}},"required":["asset_id","task_name","frequency_interval_days","next_due_date"]},"CreatePartDto":{"type":"object","properties":{"name":{"type":"string","example":"Synthetic Air Filter AF-200"},"sku":{"type":"string","example":"PART-AF-200"},"stock_quantity":{"type":"number","example":14},"min_threshold":{"type":"number","example":5,"default":5},"unit_cost":{"type":"number","example":24.5,"default":0}},"required":["name","sku","stock_quantity"]}}}}